+++ pdfminer/runlength.py
@@ -6,7 +6,7 @@
 #
 
 
-def rldecode(data):
+def rldecode(data: bytes) -> bytes:
     """
     RunLength decoder (Adobe version) implementation based on PDF Reference
     version 1.4 section 3.3.4: