+++ tests/test_highlevel_extracttext.py
@@ -34,6 +34,7 @@ test_strings = {
     "zen_of_python_corrupted.pdf": "Mai 30, 18 13:27\n\nzen_of_python.txt",
     "contrib/issue_566_test_1.pdf": "ISSUE Date：2019-4-25 Buyer：黎荣",
     "contrib/issue_566_test_2.pdf": "甲方：中国饮料有限公司（盖章）",
+    "contrib/issue-625-identity-cmap.pdf": "Termin płatności: 2021-05-03",
 }
 
 
@@ -98,6 +99,12 @@ class TestExtractText(unittest.TestCase):
         test_file = "contrib/issue_566_test_2.pdf"
         s = run_with_file(test_file)
         self.assertEqual(s.strip(), test_strings[test_file])
+
+    def test_issue_625_identity_cmap(self):
+        test_file = "contrib/issue-625-identity-cmap.pdf"
+        lines = run_with_file(test_file).splitlines()
+
+        self.assertEqual(lines[6], test_strings[test_file])
 
 
 class TestExtractPages(unittest.TestCase):